simplify use of filter trait #3376
