+++ src/Service/Behaviour/SearchQueryTrait.php
@@ -3,9 +3,27 @@
 namespace App\Service\Behaviour;
 
 use Doctrine\Common\Collections\Collection;
+use Symfony\Component\Validator\Constraints as Assert;
 
 trait SearchQueryTrait
 {
+    #[Assert\Positive(message: 'La page doit être un nombre positif')]
+    private ?int $page = 1;
+
+    public function getPage(): int
+    {
+        if ($this->page < 1) {
+            return 1;
+        }
+
+        return $this->page;
+    }
+
+    public function setPage(?int $page): void
+    {
+        $this->page = $page;
+    }
+
     public function getUrlParams(): array
     {
         $params = [];
