+++ src/Service/SearchArchivedSignalement.php
@@ -4,32 +4,13 @@ namespace App\Service;
 
 use App\Entity\Territory;
 use App\Service\Behaviour\SearchQueryTrait;
-use Symfony\Component\Validator\Constraints as Assert;
 
 class SearchArchivedSignalement
 {
-    use SearchQueryTrait {
-        getUrlParams as getUrlParamsBase;
-    }
+    use SearchQueryTrait;
 
-    #[Assert\Positive(message: 'La page doit être un nombre positif')]
-    private ?int $page = 1;
     private ?string $queryReference = null;
     private ?Territory $territory = null;
-
-    public function getPage(): int
-    {
-        if ($this->page < 1) {
-            return 1;
-        }
-
-        return $this->page;
-    }
-
-    public function setPage(?int $page): void
-    {
-        $this->page = $page;
-    }
 
     public function getQueryReference(): ?string
     {
@@ -49,10 +30,5 @@ class SearchArchivedSignalement
     public function setTerritory(?Territory $territory): void
     {
         $this->territory = $territory;
-    }
-
-    public function getUrlParams(): array
-    {
-        return $this->getUrlParamsBase();
     }
 }
