+++ src/Service/SearchTag.php
@@ -5,7 +5,6 @@ namespace App\Service;
 use App\Entity\Territory;
 use App\Entity\User;
 use App\Service\Behaviour\SearchQueryTrait;
-use Symfony\Component\Validator\Constraints as Assert;
 
 class SearchTag
 {
@@ -14,8 +13,6 @@ class SearchTag
     }
 
     private User $user;
-    #[Assert\Positive(message: 'La page doit être un nombre positif')]
-    private ?int $page = 1;
     private ?string $queryTag = null;
     private ?Territory $territory = null;
 
@@ -30,20 +27,6 @@ class SearchTag
     public function getUser(): User
     {
         return $this->user;
-    }
-
-    public function getPage(): int
-    {
-        if ($this->page < 1) {
-            return 1;
-        }
-
-        return $this->page;
-    }
-
-    public function setPage(?int $page): void
-    {
-        $this->page = $page;
     }
 
     public function getQueryTag(): ?string
