+++ src/Service/SearchTerritory.php
@@ -3,30 +3,13 @@
 namespace App\Service;
 
 use App\Service\Behaviour\SearchQueryTrait;
-use Symfony\Component\Validator\Constraints as Assert;
 
 class SearchTerritory
 {
     use SearchQueryTrait;
 
-    #[Assert\Positive(message: 'La page doit être un nombre positif')]
-    private ?int $page = 1;
     private ?string $queryName = null;
     private ?bool $isActive = null;
-
-    public function getPage(): int
-    {
-        if ($this->page < 1) {
-            return 1;
-        }
-
-        return $this->page;
-    }
-
-    public function setPage(?int $page): void
-    {
-        $this->page = $page;
-    }
 
     public function getQueryName(): ?string
     {
