+++ src/Service/SearchZone.php
@@ -6,7 +6,6 @@ use App\Entity\Enum\ZoneType;
 use App\Entity\Territory;
 use App\Entity\User;
 use App\Service\Behaviour\SearchQueryTrait;
-use Symfony\Component\Validator\Constraints as Assert;
 
 class SearchZone
 {
@@ -14,8 +13,6 @@ class SearchZone
         getUrlParams as getUrlParamsBase;
     }
     private User $user;
-    #[Assert\Positive(message: 'La page doit être un nombre positif')]
-    private ?int $page = 1;
     private ?string $queryName = null;
     private ?Territory $territory = null;
     private ?ZoneType $type = null;
@@ -31,20 +28,6 @@ class SearchZone
     public function getUser(): User
     {
         return $this->user;
-    }
-
-    public function getPage(): int
-    {
-        if ($this->page < 1) {
-            return 1;
-        }
-
-        return $this->page;
-    }
-
-    public function setPage(?int $page): void
-    {
-        $this->page = $page;
     }
 
     public function getQueryName(): ?string
